Move token bucket to separate class
